backend_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
pydantic-settings==2.5.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose==3.3.0
cryptography==43.0.1
redis==5.1.1
rq==1.16.2
httpx==0.27.2
alembic==1.13.2
